pyproject.toml
[project]
name = "vanilka-bot"
version = "1.0.0"
description = "Telegram bot for Vanilka bento cake store"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.0",
    "openai>=1.0",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "python-dotenv",
    "aiohttp",
    "beartype",
    "python-docx",
    "aiosqlite",
    "openai-whisper",
]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pyright]
typeCheckingMode = "strict"
pythonVersion = "3.12"
